CarIdentifier adding unit tests for new classes
